Refactor teams cubit and add teams repository and web services
This commit refactors the teams cubit by adding a teams repository and web services. The teams repository handles getting the teams data from the web services and converting it to a list of team models. The teams cubit now has a method to get all teams and emits a TeamsLoadedState with the list of teams. The teams web services uses Dio to make a GET request to the SportMonks API to retrieve the teams data. The AppConsts file was also updated to include the correct API endpoints.